Add files via upload
I am currently working on cleaning everything up in terms of efficiency, readability and syntax. I revised the "generateAvailableMoves" function so that it is split up into individual functions for different pieces and so that it never indexes out of bounds. I am doing something similar for the AI advantage eval.